Add testutil to reduce code duplication
