+++ internal/router/router_test.go
@@ -1,65 +1,34 @@
 package router
 
 import (
-	"net/http"
-	"net/http/httptest"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/kingcobra2468/cot/internal/router/mocks"
 	"github.com/kingcobra2468/cot/internal/service"
+	"github.com/kingcobra2468/cot/internal/testutil"
 	"github.com/stretchr/testify/mock"
 )
 
 const commandName = "test"
 const recipientNumber = "1"
 
-func NewStubServer(t *testing.T) *httptest.Server {
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/test" {
-			t.Errorf("Expected to request '/test', got: %s", r.URL.Path)
-		}
-		if r.Header.Get("Accept") != "application/json" {
-			t.Errorf("Expected Accept: application/json header, got: %s", r.Header.Get("Accept"))
-		}
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(`{"value":"fixed"}`))
-	}))
-	t.Log("started service test server")
-
-	t.Cleanup(func() {
-		server.Close()
-	})
-
-	return server
-}
-
-func NewFakeService(server *httptest.Server) *service.Service {
-	c := service.Command{}
-	c.Endpoint = "/test"
-	c.Method = "get"
-	c.Response.Type = service.PlainTextResponse
-
-	s := service.Service{}
-	s.BaseURI = server.URL
-	s.Name = commandName
-	s.Meta = map[string]*service.Command{commandName: &c}
-
-	service.AddClient(commandName, recipientNumber)
-
-	return &s
+func stopEventLoop(t *testing.T, done chan struct{}, timeout time.Duration) {
+	time.Sleep(time.Duration(timeout))
+	t.Log("stopped event loop")
+	done <- struct{}{}
 }
 
 func TestProcess(t *testing.T) {
-	server := NewStubServer(t)
-	s := NewFakeService(server)
+	server := testutil.NewStubServer(t)
+	s := testutil.NewFakeService(server, commandName, recipientNumber)
 
 	cache := *service.NewCache()
 	cache.Add(*s)
 
 	mockWorker := mocks.NewWorker(t)
-	mockWorker.EXPECT().Fetch().Return(&[]service.UserInput{service.UserInput{Name: commandName, Raw: "test"}})
+	mockWorker.EXPECT().Fetch().Return(&[]service.UserInput{{Name: commandName, Raw: "test"}})
 	mockWorker.On("LoopBack").Return(false)
 	mockWorker.On("Recipient").Return(recipientNumber)
 	mockWorker.On("Send", mock.Anything).Return(nil)
@@ -71,13 +40,9 @@ func TestProcess(t *testing.T) {
 	wg := sync.WaitGroup{}
 	wg.Add(1)
 
-	go func(done chan struct{}) {
-		time.Sleep(time.Duration(time.Second * 11))
-		t.Log("stopped event loop")
-		done <- struct{}{}
-	}(done)
+	go stopEventLoop(t, done, time.Second*11)
 
-	t.Log("start event loop")
+	t.Log("started event loop")
 	el.Start(done, &wg)
 
 	wg.Wait()
@@ -88,14 +53,14 @@ func TestProcess(t *testing.T) {
 }
 
 func TestProcess_ping(t *testing.T) {
-	server := NewStubServer(t)
-	s := NewFakeService(server)
+	server := testutil.NewStubServer(t)
+	s := testutil.NewFakeService(server, commandName, recipientNumber)
 
 	cache := *service.NewCache()
 	cache.Add(*s)
 
 	mockWorker := mocks.NewWorker(t)
-	mockWorker.EXPECT().Fetch().Return(&[]service.UserInput{service.UserInput{Name: "ping", Raw: "ping"}})
+	mockWorker.EXPECT().Fetch().Return(&[]service.UserInput{{Name: "ping", Raw: "ping"}})
 	mockWorker.EXPECT().Send("pong").Return(nil)
 	mockWorker.On("Send", mock.Anything).Return(nil)
 
@@ -106,13 +71,9 @@ func TestProcess_ping(t *testing.T) {
 	wg := sync.WaitGroup{}
 	wg.Add(1)
 
-	go func(done chan struct{}) {
-		time.Sleep(time.Duration(time.Second * 11))
-		t.Log("stopped event loop")
-		done <- struct{}{}
-	}(done)
+	go stopEventLoop(t, done, time.Second*11)
 
-	t.Log("start event loop")
+	t.Log("started event loop")
 	el.Start(done, &wg)
 
 	wg.Wait()
@@ -121,14 +82,14 @@ func TestProcess_ping(t *testing.T) {
 }
 
 func TestProcess_pong(t *testing.T) {
-	server := NewStubServer(t)
-	s := NewFakeService(server)
+	server := testutil.NewStubServer(t)
+	s := testutil.NewFakeService(server, commandName, recipientNumber)
 
 	cache := *service.NewCache()
 	cache.Add(*s)
 
 	mockWorker := mocks.NewWorker(t)
-	mockWorker.EXPECT().Fetch().Return(&[]service.UserInput{service.UserInput{Name: "pong", Raw: "pong"}})
+	mockWorker.EXPECT().Fetch().Return(&[]service.UserInput{{Name: "pong", Raw: "pong"}})
 	mockWorker.On("LoopBack").Return(true)
 
 	el := NewEventLoop(2, 2, &cache)
@@ -138,13 +99,9 @@ func TestProcess_pong(t *testing.T) {
 	wg := sync.WaitGroup{}
 	wg.Add(1)
 
-	go func(done chan struct{}) {
-		time.Sleep(time.Duration(time.Second * 11))
-		t.Log("stopped event loop")
-		done <- struct{}{}
-	}(done)
+	go stopEventLoop(t, done, time.Second*11)
 
-	t.Log("start event loop")
+	t.Log("started event loop")
 	el.Start(done, &wg)
 
 	wg.Wait()
