+++ internal/testutil/server.go
@@ -1,0 +1,27 @@
+package testutil
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"testing"
+)
+
+func NewStubServer(t *testing.T) *httptest.Server {
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path != "/test" {
+			t.Errorf("Expected to request '/test', got: %s", r.URL.Path)
+		}
+		if r.Header.Get("Accept") != "application/json" {
+			t.Errorf("Expected Accept: application/json header, got: %s", r.Header.Get("Accept"))
+		}
+		w.WriteHeader(http.StatusOK)
+		w.Write([]byte(`{"value":"fixed"}`))
+	}))
+	t.Log("started service test server")
+
+	t.Cleanup(func() {
+		server.Close()
+	})
+
+	return server
+}
